Initial DMR impl.
